software/brain/requirements.txt
# Python Dependencies — HexForge PLA Brain

# Core dependencies
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.26.4

# Web + API
fastapi==0.115.0
uvicorn==0.29.0
starlette==0.37.2
python-multipart==0.0.9
httpx==0.27.2

# AI/LLM integration (stubbed)
requests==2.31.0

# Serial communication with HID executor
pyserial==3.5

# Configuration
PyYAML==6.0.1

# Contract validation
jsonschema==4.20.0

# Utilities
python-dateutil==2.8.2
cryptography==41.0.7

# Development/Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# System monitoring
psutil==5.9.6
